LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/property-postmessage-1.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SyntaxError: The string did not match the expected pattern.

Harness Error (TIMEOUT), message = null

TIMEOUT same-origin > w => w.postMessage("", "") Test timed out
NOTRUN cross-origin > w => w.postMessage("", "")
